Renamed date util to dates
